Refactor persistence keys to use composable
